Added configuration screen for module
In addition to support for “coupon codes”, which are basically
references to sales-representatives, giving the end user an incentive
to type in a coupon code, which again creates kickback for sales rep.
Hence, basically fully fledged affiliate support for selling the system.

Also implemented quantity discount, when users are purchasing more than
5 users, according to how many tickets they purchase.

Notice, the coupon codes are checked towards online file, to calculate
discount.